-updated dashboard root layout
